+++ src/extension.ts
@@ -3,6 +3,7 @@ import { LuaAPI } from './lua_api_service';
 import { loadLocale } from './utils';
 import { JavascriptAPI } from './js_api_service';
 import { PanoramaCSS } from './panorama_css';
+import { AbilitiesDataDriven } from './abilities_data_driven';
 
 export function activate(context: vscode.ExtensionContext) {
     loadLocale();
@@ -10,6 +11,7 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push((new LuaAPI(context, true)).register());
     context.subscriptions.push((new JavascriptAPI(context)).register());
     context.subscriptions.push((new PanoramaCSS(context)).register());
+    context.subscriptions.push((new AbilitiesDataDriven(context)).register());
 }
 
 export function deactivate() {}